Test sub_pair with LaTex case, fix it partially
Separate `s:sub_pair` callback into 2 steps, first add a `s:find_pair`
callback "before" substitute, then do `s:sub_pair` with the hint of the
finding result.

This resolves the problem that, to search the "opposite" of a changed
"trigger" with `searchpairpos()` is error-prone. Because the pairs were
already broken.

- Rename some variables to make them clearer.

- Try add WarningMsg if can't find the opposite.