Fix regex matcher on multibyte with cword = cchar
When a multibyte text has cword equals to cchar (e.g., `左.`) we used to
apply a special rule that remove '' phase from phased search. The
original intention (not confidently tell) might be avoiding unnecessary
last search, but seems not really need in current environment.

Now remove the special rule, just use unified phases for multibyte case.

The test is added to "regex" suite, because the problem is discovered in
that "regex" matcher, which try to opt-out phased-search by target on
only last phases ('' or 'v') and fails if '' phase is removed.