Allow use custom function as matcher
The function must follow the interface like `matcher#xxx#test()`
functions. User can specify a function by set `matcher` to a string of
the function name prefixed with `*`, e.g., `*FunctionName`.

No checking or cache is done for the function, only call it with `*`
removed. While the dispatcher checks return type, if it is not a List,
return a "not found" list to reduce impact to outer scope.

- Add type comment to existing matcher test functions.