Start to implement "ambi_pair" internal option
ambi_pair named after "ambiguous pair".

The goal is to allow `sub_pairs` with the same character at both begin
and end sides, for example ":", ':' or |:|.

This is handled in `sub_pairs` instead of an independent option, because
it is supposed to be used with (cycle from and to) other normal pairs.

Now add `ambi_pair` if detect the `sub_pairs` has before/after set with
the same 1-character symbol.

During group search, try to find the other side with some trail methods:

- treesitter range, if available
- Vim's builtin vi" vi' text objects
- regex search