Extract function to translate group def to objects
The preferred group definition style requires a transform before it can
be actually used by cycle. The transform was handled internally, now
extract the part and expose as `cycle#parse_group()` function.

This is a start to make user define their own groups instead of stick to
`g:cycle_groups`.